packaging: Support download from GitHub release
This commit updates the convention used by versioneer
to pep440-pre [1]. Doing so allows to install the development packages
from the GitHub release. It ensures that no plus character
is included in the original package name. This one would be
switched by a dot by GitHub and would prevent successful install
of the package using:

  pip install pynwb --find-links https://github.com/NeurodataWithoutBorders/pynwb/releases/tag/latest  --no-index

[1] https://github.com/warner/python-versioneer/blob/master/details.md#how-do-i-select-a-version-style